Working grid example
Can set MAX_ROWS and MAX_COLS.
It will then create a grid and populate it with the max which can fit inside the window up to MAX_LABELS(MAX_ROWS * MAX_COLS).
When you click a grid label it will change the icon to a default icon.png. and print out the position/information about said label.
Upon resizing the individual labels will remain in their (row and column) position and will be hidden/rendered based on resizing.