lnk_to_image.py
Now when you link a clickablelabel to a .lnk file for exec path and do not include an image, it will create a folder called [row, col] and store the image in that folder. And set the image for that label to the image extracted from the .lnk file.

You can override this image by specifying an image path.

This does not load them from file yet, only shows for current session.