desktop_grid padding and name label
Made a folder for converting X -> icon (for now just lnk_to_image)

desktop_grid now has 2 global parameters (will be set in settings when i connect desktop_grid to alternativedesktop).
First is LABEL_SIZE, this is the size of the square icon image.

Second is LABEL_VERT_PAD, this is the veritical padding between rows where (DesktopIcon) name is now displayed as a label.