Valid exec_path now required
For setting executable path, it is now checked if the path to the executable is valid. For now just does not allow saving and prints to console if you attempt saving an invalid path.

Part of the reason for this is attempting to generate a thumbnail will crash the application with an invalid path.