Support for .ico external lnk files
Now has a fallback if .lnk points to a file with no icon, but uses a .ico in the same directory to set the .lnks image.

For example the discord application's .lnk points to a Update.exe which has no icon. Since it has no icon the program will search the directory for a .ico which it has. I then makes the icon from that .ico file.