Exe_to_image and improved image quality
Now uses icoextract library to extract icons from exe files. This can be run on .lnk pointers as well.

Known issue: .lnk files which point to an .exe with no icon but use a .ico file in the same directory currently do not create an image icon.

The problem i am currently having with this implementation is that if you simply copy the app.ico its rendering into the QLabel is awful. Even if you rename it to image.png (and open it and it looks perfect for a 256x256 image) it still looks absolutely awful in the QLabel. Unsure how to fix for now.