auto icon generation now checks for existing .ico file before generating
All auto image generation now runs a check for a .ico file in the execution path folder. and prioritizes this over the .lnk image or the .exe image.

Both now run extract_ico_file.py (function: has_ico_file) which returns boolean for if ico_file is found. If found this function also copies it to directory.