Command line arguments fully supported
run_program() now supports running with additional command line argments as defined in the edit menu of a desktop_icon.

This does not work for .lnk files, as I assume that the .lnk file would include all needed command line arguments. I might add support for this later.

Majority of programs do not need command line arguments. So I might separate out this line to an "advanced" tab in the desktop_grid_menu. As adding command line arguments which are undefined for the target program can cause programs to fail to launch.