warning on .lnk with command arguments
A warning message now displays if the user tries to save an icon with a .lnk exec_path AND separate command line arguments. Alternatively users can add the command line to the actual .lnk file or swap the .lnk file for the file it points to and add the command line then.