Drag and drop only changes name if name is empty
To avoid changing the name just by drag and dropping an exec_path or an image icon.

For example before if you dragged an .exe called Application.exe it would name it Application.exe and then dragged an image called Image.png into the window it would change the name of the desktop_icon to Image.png or vice versa.