Support for launching more filetypes
Launching programs now support all filetypes which have a default associated application. For example .mp4s open in your selected default video player, .txt opens in a notepad etc.

The annoying part of this commit was dealing with subprocess.Popen(command, shell=True) as this changes it to no longer throw exceptions. So warning the user that the file is not found, or no default application is set for this file type is extremely annoying.

My solution to this was to just create a pipe for stderr and check it through process.communicate(). However, in doing so it permanently links the subprocess and waits for the subprocess to terminate. (if you do not terminate it the program will just sit while "not responding").

So instead I decided to call process.communicate(timeout=0.5) which basically timesout instantly but still returns the stderr to check for errors. However, this also throws an exception upon timing out. So it needs try, except as well. Further, calling process.communicate() regardless creates a connection between the main program and the subprocess.

To fix a possible memory leak, I added process.kill() to the exception upon timeout. Which despite it's name actually just kills the connection between the main program and the subprocess (thus killing the connection and setting the subprocess free to control itself).

I both love and hate this commit....