Can now choose between multiple autogen icons
Upon auto generating an icon a window will pop up which will ask you to choose between the .ico icon and the .exe/.lnk auto generated icon.

For instance if you add an exe to Exectuable Path: which also has a .ico file in the same directory the window will pop up and you can choose the icon you wish to use.
The same exists for .lnks letting you choose between the .lnk icon and the .ico icon.