Cleanup checks before save
desktop_grid_menu.py handle_save() now is the last set of checks before saving to file.

It now holds the check for command arguments on .lnk files, and a check that the icon path exists, if not it gives the option to get a default placeholder icon, or cancel and continue editing the icon.