No longer hard coding unknown.png icon
unknown.png is now just a state of: file does not exist at icon_path
This can either a bad file path or deliberately left empty with no possible auto generated icon.

Also small fix for the warning that an icon_path is not valid, now only displays if the icon path actually contains a path (not left empty).