Added ability to delete desktop icons
In contextMenu (right click menu)  there is now the option to go into Edit or Delete. Edit does the normal desktop_grid_menu.py as was previously just right click.

Delete now sets all other fields of the item at row, column in the desktop.json to the DESKTOP_DEFAULT. effectively resetting the icon.

I did originally plan this to delete the entire row, column from the .json but I revised this to just resetting to default.
Reasoning for this is this is much less likely to encounter errors.
I enjoyed actually deleting it but the time save and storage save from a small .json file is unlikely to be of noticeable benefit.
For instance I updated render_icon() to accommodate for when I actually deleted it from the .json instead of resetting to default.
In the old version if it could not find the item in the .json then it would never call the .set_name, .set_icon_path, etc. thus never re-initializing the icon. Making the delete button essentially only work upon restart.

I left in this check for if the entry_exists(row, col) in the off chance that render_icon() is ever called for an item that does not exist in the .json.