Separators in right click menu and preparations
Right click menu now has separators between edit and the open file locations, and another separator between those and delete.

Next commit will have to deal with the new default arugments I am adding. I will be adding a line for storing a website link (with the option of left click running the file to open the link in your default browser). The left_click option will probably be set in the advanced tab, or changed in the right click menu.

I wanted to just adapt exectuable_path to be able to detect if it was a web link, but it would very likely lead to a lot of bugs with all the times I would have to check that the path is valid or website link.
Further, I believe that the setup I am going for is more ideal. Being able to store 2 "launchables" in one desktop icon with the ability to easily swap between the two means you could for instance save the website to a manual or wiki, so a left click would always open the app, then a modifier click or a quick right click menu swap and launch the associated website.