Basic website link opening
will open field "website_link" in browser.
Returns False if website_link == ""
otherwise returns True (assumed that website_link was opened in browser)

Will append 'http://' to a website_link if it does not contain it, as this ensures that it opens in a browser.

Tried using the webbrowser module and it exclusively opened things in microsoft edge (not my default browser).