run_executable FileExistsError
fixed a bug where in the case of an .lnk it would not check if the file actually existed, thus would not display a message box to the user.

it now runs a check before attempting to run for if the file at path exists. Since this split it also clarifies the reasoning for the "is not recognized as an internal or external command" error to only be a raised as a result of no default application set to open file type. So I have adjusted the two QMessageBox warnings to clarify their reasoning.