Clicking out of right click menu no longer launches
There is now a delay on clicking to cancel out a right click menu.

For instance if you right clicked on something and then clicked on another icon just to try and get rid of the right click menu before it would launch said icon.
Now upon closing the icon there is a 100ms delay before a click can be used to launch an icon.
Also has a backup method of resetting CONTEXT_OPEN to false if you left click any icon (backup in case some bug gets the user stuck in CONTEXT_OPEN = True)

Annoying implication having to use globals and timer because a ClickableLabel variable would only work for the icon you clicked (thus needs global) and a timer is needed because context_menu_closed() will be called before mousePressEvent()