Hover over name now shows full name tooltip
Hovering over the name label of a desktop_icon for more than a seconds now produces a tooltip which displays the full desktop_icon.name.

This is mostly useful for long names with small icons. Because at least for now, the larger the icons the larger the label for desktop_icon.name