Added file path buttons for icon and executable
Added buttons to the desktop_grid_menu next to the Icon Path field and the Executable Path, to allow the user to find the file through explorer and set either field to it.

Drag/drop of non image files still go to Executable, and you can still drag and drop an image file to set Icon Path.