Basic support for gif icons
When you set a desktop icon to a .gif file it now attempts to play the gif as the icon.

Problem with this seems to be that it does not actually show the correct ratio size so it just plays the LABEL_SIZE pixels worth of what it sees in that gif. For example in a 100x100p gif if the LABEL_SIZE is 20, then it will play the middle 20x20 pixels of the gif (not the full 100x100).

So while it works for now you need the gif icon to match the LABEL_SIZE.