Basic .mp4 video background support
Now attempts to set a background (for now hard coded at the file path where you run this -> "background.mp4").

The video will scale to attempt to match the Grid Widget's resolution, For instance If you keep the Grid Widget's size at a 16:9 format and have a 16:9 background.mp4, then it will fit exactly into the window. But if you resize the window so it is no longer 16:9 (for instance make it taller, shorter, fatter, or skinnier) It will add paintEvent's background colored bars to the rest of the window to fill.

I have set the default size of the window to 1760x990 (16:9 resolution) Because both the window bar and the windows taskbar will prevent true 16:9 resolution = 1920x1080 on 1080p monitors.

However, my full idea of how this will be implemented is that people will generally use it maximized to fit their full screen. So I think I will probably have to try and find a way to set the background to always fill the entire window if possible and then you can select to essentially cut off part of the top or the bottom etc. to make up for the taskbar and window title. It sounds like a pain but probably doable.