favicon get now directly based on google link api
Removed a lengthy requests.get() sequence for getting an icon from a website.
For most sites it was the same as if not slightly slower and lower quality than the google api (simply a get request to the website: "https://www.google.com/s2/favicons?domain={url}&sz={ICON_SIZE}". And at the same time trying to get it from a website that is not responding (no responses or just bad website name etc.) would cause a very long pause in the program while it awaited the server to respond, up to 15 seconds to reach max retries.
I have instead just focused on believing that the google api will generally work in almost all cases (even works on some obscure websites I have tried), and just default down to browser icon should it not be found through the google API.