Complete rewrite for minimize/maximize hotkey
Completely redesigned the way the keybind was handled. Old design caused massive problems with the way things were handled in the grid object.
It is now mostly within its own hotkey_handler.py class. With the exception of toggle_window_state() and change_event() which are in the desktop.py (parent class)