Minimizing program now pauses video background.
Now pauses the video when you minimize the program through the keybind. I will probably make this a setting to pause or not. But for now it is by default pause on minimizing through keybind. Would probably have to override the base titlebar minimize if I wanted both to pause the video as that does not pause it.

Secondly, pausing the video does not seem to use any less resources in the background than an unpaused video (so long as it is minimized). So this is not quite an optimization setting more of a "Do you want the background video to pause or not" setting.