Full support for Background Source setting
Can now set the specifics for a background source.
First found: Will load a video if it exists, if it doesn't will load the background image if it exists, if not then loads neither.
Both: Loads both (video overtop image). NOT RECOMMENDED as it causes lag but the option is there.
Video only: Only try and load video (if it doesn't exist load nothing)
Image only: Only try and load image (if it doesn't exist load nothing)
None: don't load either and keep the background as a default fill color.

Background video and image are still hard coded to "background.mp4" and "background.png" for now, will update to setting path soon.