Swap folder and icon_path on swap icons
Will now call swap_folders() which will swap the [row1, col1] and the [row2, col2] folders in the data directory. (i.e. swap [4, 10] folder with [4, 11] folder when you drag and drop swap them).

Also calls update_folder in config which if the path of the icon_path in config starts with the data directory (AppData\Roaming\AlternativeDesktop\data\) will update the icon_path to use the new [row, col] file it has swapped.