No longer allow icon drops on self.
dropEvent() now returns if icon dropped is the same as the icon dropped on (drag and drop in the same location). Didn't change anything when there was no file swapping etc. but now that files are swapped it starts causing errors.

Also added the ability to create a _temp folder with a unique name. I only had problems with this during testing before things were fully implemented and the safeguard for drag and drop on same icon. But added a way to still continue and give a warning instead. Please let me know if you get this warning and any steps to reproduce if you can reproduce it.