Now gets default icon from default associated program
Can now generate a default icon from the default associated program. For instance .txt associated with notepad++ then it gets the icon for notepad++.
Basically the icon that the file has in windows file explorer is now loaded and saved as an option. Note this is only for non lnk, exe, url files as those are already covered through other thumbnail_gen files.