V0.1.0 prep work
Changed the format of AlternativeDesktop.py to support the new installer formatting.

Before I was including the other .py files but the AlternativeDesktop.exe was not actually even using those included files. Thus any changes to them were pointless.

The direct imports for AlternativeDesktop.py are now loaded in as modules from the current directory THEN defaults to the ones bundled in the .exe if they do not exist. Therefore the included .py files in the installer (when I release new builds) will be editable without having to rebuild the .exe (can edit in notepad without needing an entire python environment and deep coding knowledge).
For instance back in the day when I played python based games they had this setup and I used to edit certain functions or variables for *cough* totally non-cheating means *cough*.

This setup is also just a lot more convenient for debugging on my non-python pc setup, allowing me to just edit the files there to debug.