Removed pylnk3
All my homies hate pylnk3.

On a real note though, this was one of the most annoying libraries I have ever worked with. Constantly returning None for base path/other paths seemingly randomly. Baking into the .exe with pyinstaller making it appear that it didn't install in _internals. Trying to trace this down and turns out its in the .exe itself. Then trying to special case it only to realize that the paths keep fucking up and can't be resolved on any location other than base Desktop. It is a bit out of date so I assume some of these problems have popped up over the years.

Also moved .ico to base folder.