Basic logging setup.
Added a basic logging setup to output logs to:
/logs/alternative_desktop_TIMESTAMP.log

NOTABLE PROBLEMS:
The library icoextract seems to also use the base python logger. Which seems to mess with the ability to write to the log file. Thus when creating the log file from desktop.py IT MUST be imported AFTER setup_logging is called (Inside OverlayWidget()). Otherwise it will overwrite the logger and stop it from writing to the logger created by our program.