Fixed possible crash if toggle overlay hotkey not an actual hotkey.
Fixed a problem of instantly crashing the program if the toggle_overlay_keybind in settings failed to be a correct keybind as recognized by pynput keyboard. It now tries to assigning the keybind and reverts to a default of Alt+d if it is not a valid keybind.

Also added new display_warning.py to better organize some warnings.