default_icon_to_image now retries and has a fallback
default_icon_to_image now tries 5 times to get a valid hIcon response in testing with it it seemed to return a valid response ~90-95% of the time. Thus 5 retries should very nearly catch all cases and almost guarantee a valid response. In testing nearly 1500 attempts, only 4 made it to attempt 3, and none made it to attempt 4/5.
In the case that it does not receive a valid hIcon after 5 attempts it will fallback and set the icon to the base unknown.png

Also added logging to better keep track of default_icon_to_image.py