Updater now has confirmation message
Added a QMessageBox on launch telling the user that a new version has been downloaded and allow them to install it or not. (only tries to update at the moment if you have "Update on Launch" setting turned on and only during launch).
This is to give the user some idea what the program is going to install an update, rather than just nothing happening then windows asking for permission to install the program. (sketchy)

Because a QMessageBox requries a QApplication I have also changed it such that we initiate the QApplication through desktop.create_app() before calling for updater such that it can use that QApplication instead of creating its own. (Deleting a QApplication and making a new one is hard/not intended to be done so you are supposed to re-use it).