Reformatted desktop icon previews
preview_icon_path() now runs any time the icon_path Line edit is changed. It attempts to preview the desktop icon with the path in the line edit. If it exists it will preview the icon, if it does not exist it will show the unknown.png (? icon). Much better to set it to a .textchanged.connect function than to try and chase down all the different versions and add the same code to them.

Also removed a literal 1 line function which was only called a few times and replaced it with the line.