Minimize to system tray upon close.
Now when you close the window, it instead minimizes to system tray. Keybind to bring up the window will still respond. So if you accidentally X close or close the program instead of keybind minimizing it, you can still bring it back up normally without having to relaunch the program.

I will probably add a setting menu for this, defaulted to minimize to tray on close.