Optimized keybind restore and simplified variables
Optimized it such that when restoring a Maximized window it does not run self.showNormal() first, thus making it quicker and not flash a small window then a large window upon restoring by keybind. The only time it runs self.showNormal() is the first run where it seems to have to run self.showNormal() in order to properly fill the component, but for every restore after that it is quicker.

Greatly simplified the two class variables for first restore through keybind.