Improved WindowState management and restoring window
Fixed some problems with restoring the window after closing the window from taskbar while the window was minimized.
Removed self.previous_state jumbled mess and now just keep track of the two states that matter. NoneState (open but not minimized Also referred to as Normal, and Maximized).
The state is easy to identify as follows:
self.is_maximized == True == Maximized
self.is_maximized == False == NoneState/Normal
This is tracked through changeEvent which handles WindowStateChange and only updates on these two states since these are the only two we need to keep a last seen state of.

Also added that when you close out of the window while the window is minimized it calls to showNormal or showMaximized immediately before hiding. This is required for the restore to actually place the window on top, as if you hide it while minimized it will restore it in an open non-minimized window but it will not appear on top of your other open windows. It simply opens it in the background and would then require several more restores to show on top.

Also updated show_window() (the taskbar restore contextmenu option) to use the same toggle_window_state call as the keybind. Thus mirroring the same functionality although it could be overriden if for instance I want it to always open full screen on restore after closing to taskbar.