Window restore optimization
Returned the optimization from earlier versions where it only calls showNormal() for the first restore of a maximized window. Once this is done every restore after is correctly Maximized. But if you never call showNormal() then the Maximized restore maximizes the window but the components inside do not take up the full window.