Small settings_menu optimizations
Because this deals so heavily with all the settings, it is more efficient to just load all settings into a self variable and run .get() on that rather than running Settings.py's get_setting() function.

get_setting() is more for like one off I just need this setting, as such it loads directly from the file (load_settings()) every time it is run to ensure an up to date status.