Slightly optimized Settings
Looking over the settings during the last commit I realized that there's really not that much of an importance to reload settings all the time. Settings only get changed either on launch in check_for_updates(), save_settings(DEFAULT_SETTINGS) for a new install, or when save_settings() is called.

Therefore you can just make sure settings are loaded on launch, then call load_settings() every time save_settings() is called and a global variable will remain up to date and not require calling load_settings() every time you want to get a setting.

For settings_menu.py since the settings are accessed so much doing the 1 time load_settings() is now more of a preference than an optimization. But for all the one off get_setting() calls this should be a very small optimization.