Fix resizing height getting stuck.
I forgot the jankiness of this not wanting to un-render a row once it had rendered it. The only real solution to this was to change their fixed sizes such that they go under a certain size and can collide enough to un-render a row. I think I might revisit the way that these are displayed eventually.