Unmaximizing now shows a better aspect ratio window
Upon unmaximizing the program used to be this awkward tall and skinny shape which tended to be too tall to fit on the screen where it appeared in the middle. I have fixed this to try and match the aspect ratio of the monitor and draw it in the center of the screen upon first resize. After that it maintains the aspect ratio you last had upon unmaximizing.