Slight fix for multiple monitors
Now resets first_resize upon changing monitors, as such upon un-maximizing on a new monitor it should attempt to redraw at 75% width, 75% height of the monitor. It will also attempt to place it on the correct screen. With the last commit it would get the 75%s right (only the first time) but place it on the main monitor.