Fix some resizing bugs for multiple monitors
When changeEvent is called to normal windowed mode for the first time on a monitor, it now resizes it bounded to the max height and width of the monitor, before trying to scale it to 75% of screen resolution.

This fixes some bugs I noticed when taking an abnormally tall window to a relatively short resolution monitor (For me was easy to notice with 1080x1920 monitor to 1920x1080 monitor). For instance it would resize to 75% then when adding labels would cause another resizeEvent which I could not track down.

While it still can increase the height slightly beyond 75%, I have not seen any situations where the height of the window is larger than the height of the screen. As height is not 100% set in stone, I have moved the window to appear just 10 pixels below the top of the screen. This should mean even the worst case it should at least fit on the screen and now go far below the screen like the old version was doing.