Patch notes popup proof of concept
New settings.json variable: "show_patch_notes" this would be changed to True every update. When true upon launch, shows a popup window with the current patches patch notes and the ability to view older patch notes.

For this patch it does not turn back show_patch_notes to False after showing the window. This is just for testing its functionality.