Small updates for display_warning returns
Some uses of these display_warnings were explicitly looking for QMessageBox.Cancel to be returned.
While this seemed to work, at least for me, when X closing out of the window it would still return QMessageBox.Cancel it does not appear that this is supposed to be used this way as it should be returning QMessageBox.Rejected. I assume there is some equivalency backend that can assume some logic to say Rejected ~~ Cancel. But this may not be 100%.

Even further it is just 100% better code to look for the one affirmation button and if not then assume cancel/rejected.