Keybind/restore now brings to top if not minimized
When the window is still open but not the active window it will now bring it back up and into focus.

Before it would minimize it, then require another keypress to bring it back to top. It is much more likely that when pressing the keybind the user would assume it would bring it to top rather than just minimize it when not in focus.