Fixed black screen while showing as maximized
Old line of code which I was using when trying to sort out the multiple monitors/entering windowed mode from maximized not maintaining an aspect ratio bounded by the size of the current monitor.

Basically this old line of code was causing 2 changeEvents() thus calling it to redraw the screen, instead of just showMaximized().