Full support keybind_minimize setting
Fully implemented the "When in focus Keybind:" setting with 2 options.
1: Minimize window, this is the default and what the program has used up till now. Just pressing the keybind when the window is in focus will minimize it to taskbar.

2: Hide Window, This is new and will hide the window with the only way to restore it being through pressing the toggle overlay keybind OR through the windows system tray (right click -> restore). Useful for removing the window from taskbar and only showing the window when it is wanted/called.

The "When in focus Keybind" Hide window setting is separate from the "on closing the program" setting. "On closing the program" explicitly deals with X close and right click -> close. While "When in focus keybind" deals explicitly with in focus keybind button press.