Fixed a bug that caused f10/f11/f12 keybinds to crash
Updated add_angle_brackets to search first for larger keys before searching for smaller keys.

For example the old code when passed f10 would return <f1>0 because it found the key f1 Then put angle brackets around it and did nothing for the 0 as the 0 is not a key.

Now when searching for keys it first searches for longer keys like f10, then if it doesn't find it searches for shorter keys like f1 etc.. So it returns <f10>