Display warning when keybind is not fully supported.
When setting the keybind to a numpad keybind which is not num 0 - num 9, or the key "/" it now displays a warning that the keybind is not fully supported.

These keys are the only major issue keys that I have found in basic testing. Some of them might be fixed to be fully supported but at the current state a warning on setting the keys to these should be helpful for a user.