Simplify and use Qt.Key references
Simplified it down to only use numpad_mapping on the ones that need more than just "num ". Specifically all the ones that instead send a special key Not their base version.

For instance shift+numpad 7 sends Home key so you have to intercept it and remap it to "shift+num 7" etc.