Reworked warning for unsupported keybinding
Since almost all problems with numpad keybinding is fixed. I changed the display_keybind_not_supported() warning to only show if the scan code returned is in the range of major error (57000+) which is what returns for '/' 'num /' and 'menu' I have not found any other instances which cause errors like this. If you find any let me know.