Fixed numlock on/off differences.
Fixed some problems where setting a keybind with numlock on vs the same keys with numlock off would change the keybinding. For instance:

Numlock off, shift + num 7 would set the keybind to "shift+shift+num 7"

Numlock off, num 7 would set keybind to "shift+num 7"

Basically when keys like "Home", "Left","Insert" etc. were called for keypress it would set the keybind to "shift+" version assuming that numlock was on. As Numlock on returns those same "Home", "Left", "Insert" etc.

This is now modified to keep track of the shift being pressed independently thus allowing the same exact key mapping setting and use whether numlock is on or off.

Most of the annoyance with this commit is due to how keyboard firmware itself decides what type of keys are sent to the OS, thus can remove modifiers like shift when converting from numpad 7 to Home etc. Instead of just sending the key with its modifiers for the OS to decide what to do with it.