Final V0.1.001 release
This is the build that was compiled and uploaded as release. (30min after original posting)
Found out while updating on a different setup that the program running while installing causes problems, where it won't close, and requires a hard close by the user to update. I changed it for now to just full close, and assume that it installs correctly/gives an error for the user to report.

Also changed the default X close to terminate instead of minimize to tray, It's nice if you know its enabled, but should not be default.