Updater now detects if update already downloaded
Updater now checks if latest version already downloaded in temp folder. If it is then it gives a menu to install that downloaded version, or the option to redownload and install it.

Also reformatted updater to use more global values for stuff. There was too much passing around of the same variables that it got kind of confusing as to what had which and how to get something like latest_version a few functions down from when it was last in scope. So now things like the FULL_LOCAL_PATH and LATEST_VERSION are also global variables which get set as soon as it is found that there is a new version and do not need to be updated ever.