new_grid fixed grid and only draws in window
Fixed an issue that had the grid items appear in [column, row] instead of what we should be doing [row, col].

Also added a set_color function which changes the color of the icon (just used for now to test calling methods on a DesktopIcon from DesktopGrid). And ensuring that DesktopIcon's mousedoubleclickevent() works.

Further DesktopIcons are not stored in a self.desktop_icons. And can be referred to/call functions via: self.desktop_icons[row][col].set_color("red") for example.