new_grid slight optimization
Resizing the window with a large amount of icons could be a bit laggy. I have gone with an approach that will prioritize resizing the window. Then redraw after a short delay. This makes it a bit better with resizing actually following the cursor, at the expense of some of the icons moving, before being redrawn upon stopping/ every so often while dragging.