new_grid slightly improve resize performance.
TLDR: very slight performance increase during resizing. (only in new_grid, not in live or installation version).

Went on this whole thing trying to track down why resizing with a nearly full window can cause some lag. Thinking it was because of the way I iterated over every single self.desktop_icon and set its visibility every time it was resized, I improved the setVisible() to only call on the desktop_icons which have either been added to scope of window or removed from scope of window.

Then I realized this is mostly just down to how the graphicsview draws the screen and deals with resizing. This is probably the extreme testing though because it is unlikely that a user would actually set 400 icons on the desktop, and then complain that spaming circles while resizing the window was getting a bit laggy....

Actual use cases where blank icons are not rendered a user should be able to spam circles while resizing just a little bit better thanks to this change though :)