new_grid "background.png" and Text outline
Reminder to run: "py AlternativeDesktop.py prototype" In order to see the prototype "new_grid"

Added a direct "background.png" (INISDE MAIN FOLDER) attempt to set the background. (mostly just to confirm it works as expected).

Added a text outline to the desktopIcon names. For now it is just a pre-defined 4px black outline. This is something that was pretty much impossible in the desktop_grid.