new_grid multi line names now split long words.
Reminder to run: "py AlternativeDesktop.py prototype" In order to see the prototype "new_grid"

Previously it would only split off at the end of a word thus a long word would go on far beyond the edge of the icon_size.

This now splits words and makes sure to fit within the horizontal barrier of the width of the desktopIcon. It also limits the multi line text to 3 lines adding a "..." at the end should it be longer than 3 lines worth for a name.

Will probably work on a similar system to get a max lines that fit in the DesktopIcon name's allocated space (VERTICAL_PADDING) such that if there is more vertical padding it will allow more lines and less lines for less padding etc.
Ideally will have settings for fonts/font sizes as well so will need a lot of work, and much of it might come after fully replacing the desktop_grid() in release with the prototype new_grid()