new_grid added context menu upon right click
Reminder to run: "py AlternativeDesktop.py prototype" In order to see the prototype "new_grid"

Added a context menu with 1 action "Edit Icon" which is not currently hooked up due to a lot of problems that will need to be solved. Like rewriting part of desktop_grid_menu, since QGraphicsViews cannot be set as a parent unlike the old Labels.
Thus desktop_grid_menu needs to be rewritten to not refer to self.parent() and still needs a way to communicate back to the QGraphicsItem for things like updating the icon etc.. Doing so will also mean breaking the usage of Desktop_grid which means a full commit to new_grid prototype. Which will be coming.

This commit also replaces the self.media_player from QGraphicsView with a global MEDIA_PLAYER. This is because in desktop_grid we could just call self.parent().media_player.pause() but when using a QGraphicsItem it cannot have a parent like the Labels could thus needs a reference to the media_player with the simplest implementation just using a global reference.