new_grid fixed resizing restarting video background
Reminder to run: "py AlternativeDesktop.py prototype" In order to see the prototype "new_grid"

Video source now only gets changed in render_bg() if the BACKGROUND_VIDEO global variable has changed. The problem that was restarting the video every time it resized was that it was loading the video_source like a brand new source every time you resized. Instead of only loading it when it has been changed.