Migration new_grid to desktop_grid
Completely replacing the old desktop_grid QWidget/QGridwidget with the new QGraphicsView desktop prototype.

Up till now you could have used either the release installer version or the most up to date version but following this commit you should either migrate to the installer version if you were running it from the code base or just accept that some things will not work for the time being but will be restored over the next few days/weeks. It is just much better to replace it early and restore everything than to try and maintain 2 separate versions while still making large changes.

It is not fully caught up to functionality as the old one. A lot of work still needs to be done. But the improvements over the old desktop_grid is immense. It launches faster, uses less resources, reloads the screen faster, and has room for a ton of improvements over the old desktop_grid. Changing fonts, font size, text outlines. None of the funkiness of the window size changing itself because it thinks it needs to add another line and can only add another line by changing the entire window size....