Edit icon/calling desktop_grid_menu
Was struggling yesterday dealing with a way to communicate between the QGraphicsItem (desktopIcon) and the desktop_grid_menu since QGraphicsItem cannot be a parent. However found out that I can just call the QGraphicsView itself as a parent and then call back to the desktopIcon itself. This will be how we can easily migrate over functionality from desktop_grid_menu to the new desktop_grid.

NOT everything is working for it, for example an auto generated icon will cause an error, and saving in some situations can cause an error as well. This is mostly just an early commit to show how I can have desktop_grid_menu communicate back to the desktopIcon.